engine/requirements.txt
MetaTrader5
pandas
numpy
ta-lib
python-socketio
requests
python-dotenv
